+++ src/components/Fieldset/FieldsetLegend.tsx
@@ -1,0 +1,71 @@
+import classNames from "classnames";
+import PropTypes from "prop-types";
+import React from "react";
+
+import "lbh-frontend/lbh/components/lbh-fieldset/_fieldset.scss";
+
+/**
+ * The proptypes for {@link FieldsetLegend}.
+ */
+export interface FieldsetLegendProps {
+  id?: string;
+  className?: string;
+  /**
+   * Whether the legend also acts as the heading for the page. Places the legend
+   * children inside a `h1` if set to `true`.
+   */
+  isPageHeading?: boolean;
+  children: React.ReactNode;
+}
+
+/**
+ * A component to set a legend for a {@link Fieldset} by passing
+ * in via the {@link FieldsetProps.legend} prop.
+ *
+ * ```ts
+ * // Creating a fieldset with a legend
+ * <Fieldset
+ *   id="fieldsetId"
+ *   className="fieldsetClass"
+ *   describedBy="fieldsetDescribedBy"
+ *   role="fieldsetRole"
+ *   legend={
+ *     <FieldsetLegend id="legendId" className="legendClass" isPageHeading>
+ *       children
+ *     </FieldsetLegend>
+ *   }
+ * >
+ *   children
+ * </Fieldset>
+ * ```
+ */
+export const FieldsetLegend: React.FunctionComponent<FieldsetLegendProps> = (
+  props: FieldsetLegendProps
+): JSX.Element => {
+  const { id, className, isPageHeading, children } = props;
+  return (
+    <legend
+      id={id}
+      className={classNames(
+        "govuk-fieldset__legend",
+        "lbh-fieldset__legend",
+        className
+      )}
+    >
+      {isPageHeading ? (
+        <h1 className="govuk-fieldset__heading lbh-fieldset__heading">
+          {children}
+        </h1>
+      ) : (
+        children
+      )}
+    </legend>
+  );
+};
+
+FieldsetLegend.propTypes = {
+  id: PropTypes.string,
+  className: PropTypes.string,
+  isPageHeading: PropTypes.bool,
+  children: PropTypes.node.isRequired
+};